Fix: Fixing Coding Structure
